♻️move timer to tools
while it wraps a browser API, it is mainly used to control the execution flow
