+++ packages/core/src/tools/timer.spec.ts
@@ -1,7 +1,7 @@
 import { mockClock, stubZoneJs } from '../../test'
 import type { Clock } from '../../test'
-import { noop } from '../tools/utils'
-import { resetMonitor, startMonitorErrorCollection } from '../tools/monitor'
+import { noop } from './utils'
+import { resetMonitor, startMonitorErrorCollection } from './monitor'
 import { setTimeout, clearTimeout, setInterval, clearInterval } from './timer'
 ;[
   {
