+++ packages/core/src/transport/batch.ts
@@ -2,7 +2,7 @@ import { display } from '../tools/display'
 import type { Context } from '../tools/context'
 import { computeBytesCount, jsonStringify, objectValues } from '../tools/utils'
 import { Observable } from '../tools/observable'
-import { setTimeout } from '../browser/timer'
+import { setTimeout } from '../tools/timer'
 import type { PageExitEvent } from '../browser/pageExitObservable'
 import type { HttpRequest } from './httpRequest'
